Generate Kotlin Snapshot Files for Incremental Compilation
Soong now creates a rule to generate .snapshot.bin files for every
jar. It then threads through a map of jars to snapshots to the
kotlin compilation rule where every jar that is used also has its
corresponding snapshot added as a dependency.

These snapshots represent kotlin's view of the abi for a jar. It uses
this snapshot to determine if the dependencies of an incremental
build have changed. Without this, it is possible to change a
dependency in such a way that incremental  compilation succeeds, but a
runtime failure due to a signature mismatch occurs. The easiest
way to repro this is by adding a parameter to a method in a kotlin
dependency that has a default value set. No change in any other
source files is needed for incremental compilation to succeed, but any
clients of that dependency will fail to execute properly.

If a snapshot is missing, this results in a ninja error, as the
snapshot will be added as a dependency but there will be no rule
to generate it.

Excess rules for snapshots that are never depended upon are likely
written, but the usage of a jar-snapshot map ensures that they
don't interfere with the build, and that snapshot order is consistent,
dependent upon the jar order.

Fixes: 419452540
Test: m
Test: Custom mini app written to repro the bug.
Change-Id: Ic6cdb8bae1712bc75420e59f73b0c0f09f5df13c